Fixed visible runs in Liverpool (#6)
* started writing code to check runs

* visible_cards needed definition modification in Liverpool; 
In Liverpool (but not HandAndFoot) changed structure of visible_cards so that it is now a list containing a single dictionary, instead of a dictionary for each player.  That single dictionary is the the one with the most cards played of all the dictionaries received from the player channels.  This allows you to preserve cards played by a player who later drops out. AND to preserve the ordering of runs, which should simplify assigning wilds.