Runs rules and sorting (#8)
Lots of rule checking. However, there is now a bug that appears if prepare a wild for a RUN and then hit play prepared cards.

Significant changes-- Cards now have property card.tempnumber so that can track assigned values of wilds, and when Aces are high. Server code does NOT use this property, as it is not passed between players and server.

processRuns is a new method that puts cards in order (including unassigned jokers). It simultaneously does rules checking.
Does not yet handle jokers that can be either High or Low. Need to sort out whether this should be called before canPlayGroup or inside it. (currently does both)..  Note canMeld assumes that it will be called within canPlayGroup.
However, need to make certain that processed cards are what ends up transmitted to the server.