Add Liverpool to this fork of RemoteCardGames. (#12)
Modifying Hand and Foot code so that it supports Liverpool.  

In this version-
Verified:
- HandAndFoot still works.
-pyinstaller still works (creates executable)

Player dropping out still represents a challenge in Liverpool, but can now restart game in any round.

**Shared board -- can play on other players:
* Prepare cards button per group (#1)
  - Rather than single prepare_cards_btn now have assign_cards_btns.    (Also created method to change text within a button, but didn't end up using it).

* visible_cards needed definition modification in Liverpool; 
In Liverpool (but not HandAndFoot) changed structure of visible_cards so that it is now a list containing a single dictionary, instead of a dictionary for each player.  That single dictionary is the the one with the most cards played of all the dictionaries received from the player channels.  This allows you to preserve cards played by a player who later drops out. AND to preserve the ordering of runs, which should simplify assigning wilds.   Instead of keying on card.number, key on tuple = player#, group#.   

*. played_cards in HandAndFoot can be managed  locally, but it cannot be local in Liverpool.
* Mechanics a bit different because you designate the group cards go to explicitly upon preparing them (each group has a different button) but you don't designate wilds value until after you hit play prepared cards button.


**Buying Option:
* Pick up pile is working, as is buying cards option.
  - Liverpool doesn't require that top card be played--modifications made to support this.

* Concluded buyOption had to be controlled by server.
Got server to send announcement to eligible players that they could buy top card, so server lets eligible clients know when buying opportunity occurs.  Clients let server know if they want the card.    Gives purchased card and top card to player who buys it.  Server also  announces if someone buys the card or picks up the pile (or top discard).
Reset want_card so that don't accidentally purchase future discard if wanted earlier one.


Ran into 2 related problems, both traceable to the fact that Liverpool allows you to play on others sets and groups.

**Also have runs:
* Implemented more sorting options: by suit and Aces high / low sorting options.
* To do rule checking have to preserve  values assigned to Jokers.
-> gave card another attribute (not used by server) -- tempnumber.  Need this to preserve assigments for Jokers, and whether Aces are high or low. 
* Created method proccessRuns  in liverpool.py.  Puts run in order and  does some of the rule checking.  

** Other variations in rules:
Modified canMeld
Moved message that appears at start and between rounds to ruleset (both HandAndFoot and Liverpool).
etc.