Server pick game (#29)
Main update is that now the client gets the name of the ruleset from the server.  This required changes to RunClient.py,
Controller.py, Clientstate.py,  GameServer.py and PlayerChannel.py.

Minor change:  hostinfo and port now entered separately.
Did this so that players would not need to type in host name when we switched to another port#, but could continue to use default.  

Also verified the if/while statements used to set player_index in RunClient.py are needed (while loop goes through a couple of times).